Finished moving tests from test_atomgroup
